Put firehose specific stuff in its own package.
